Added User Module using lazy-loading and removed driver component
