+++ src/app/_modules/user/user-routing.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { UserComponent } from './user.component';
+
+const routes: Routes = [{ path: '', component: UserComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class UserRoutingModule { }
