+++ src/app/_modules/user/user.component.html
@@ -4,7 +4,7 @@
             <mat-card class="box">
                 <img  src="/assets/logo.svg" width="200px" />
                 <mat-card-title class="text-center">
-                    <h3 class="mat-h3 ">Client Profile</h3>
+                    <h3 class="mat-h3 ">Profile</h3>
                 </mat-card-title>
                 <mat-card-content>
                     <p>Update your personal information by clicking the edit button below</p>
