+++ src/app/_modules/user/user.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-user',
+  templateUrl: './user.component.html',
+  styleUrls: ['./user.component.css']
+})
+export class UserComponent implements OnInit {
+
+  edit = false;
+  
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
