+++ src/app/_modules/user/user.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { AngularMaterialModule } from '../../angular-material.module';
+import { UserRoutingModule } from './user-routing.module';
+import { UserComponent } from './user.component';
+
+
+@NgModule({
+  declarations: [UserComponent],
+  imports: [
+    CommonModule,
+    UserRoutingModule,
+    AngularMaterialModule
+  ]
+})
+export class UserModule { }
