+++ src/app/app-routing.module.ts
@@ -1,5 +1,3 @@
-import { DriverProfileComponent } from './user/driver-profile/driver-profile.component';
-import { UserProfileComponent } from './user/user-profile/user-profile.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { LoginComponent } from './components/auth/login/login.component';
@@ -9,12 +7,11 @@ import { GuestsigninComponent } from './components/auth/guestsignin/guestsignin.
 
 const routes: Routes = [
   { path: '', pathMatch: 'full', redirectTo: 'register' },
-  { path: 'user', component: ForgotpasswordComponent },
+  { path: 'forgotpassword', component: ForgotpasswordComponent },
   { path: 'login', component: LoginComponent },
   { path: 'register', component:RegisterComponent  },
   { path: 'guest', component:GuestsigninComponent },
-  { path: 'profile/client', component: UserProfileComponent },
-  {path: 'profile/driver', component: DriverProfileComponent}
+  { path: 'user', loadChildren: () => import('./_modules/user/user.module').then(m => m.UserModule) }
 ];
 
 @NgModule({
