--- src/app/user/driver-profile/driver-profile.component.ts
@@ -1,17 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-driver-profile',
-  templateUrl: './driver-profile.component.html',
-  styleUrls: ['./driver-profile.component.css']
-})
-export class DriverProfileComponent implements OnInit {
-
-  edit = false;
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
