--- src/app/user/user-profile/user-profile.component.ts
@@ -1,17 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-user-profile',
-  templateUrl: './user-profile.component.html',
-  styleUrls: ['./user-profile.component.css']
-})
-export class UserProfileComponent implements OnInit {
-
-  edit = false;
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
